credentials: per-provider plugin files
config/plugins/credentials/ used to be five cross-cut files
(credentials.go, slack_hitl.go, secret_slots.go, env_pushdown.go,
oauth_flows.go) — every method for a single provider scattered
across them. Split per-provider: each credential lives in its own
file (slack.go, postgres.go, github.go, …) carrying the struct plus
every interface method (HTTPCredentialRuntime / TLSCredentialRuntime
/ PostgresAuthCredential / HITLNotifier / SecretSlotsProvider /
EnvPushdownProvider / OAuthFlowProvider) plus the init() that
registers the plugin.

util.go retains the cross-cutting helpers (newer / passthrough /
ensureBeta / placeholder constants / emptyEmit).

Files:
  bearer_token.go cookie_token.go header_token.go mtls.go
  postgres.go    clickhouse.go    anthropic_manual_key.go
  anthropic_oauth.go openai_codex.go github.go notion.go
  slack.go       telegram.go      gemini.go      aws_eks.go
  util.go

Behaviour unchanged — each Plugin.Register call comes out byte-
identical to the previous centralized init().
